Separação do Quiz em 3 Telas
