+++ package.json
@@ -18,6 +18,7 @@
     "react-dom": "16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
     "react-native-elements": "1.2.7",
+    "react-native-expo-viewport-units": "0.0.8",
     "react-native-gesture-handler": "1.6.1",
     "react-native-paper": "3.9.0",
     "react-native-reanimated": "1.7.0",
